search and sort algorithm implemented. Binary, Interpolation search; Bubble and insertion sort;
