+++ DSA/DSA.BLL/DSA.BLL.csproj
@@ -51,7 +51,11 @@
     <Compile Include="DataStructures\Queue\VectorQueue.cs" />
     <Compile Include="DataStructures\Queue\ArrayQueue.cs" />
     <Compile Include="DataStructures\Queue\IQueue.cs" />
+    <Compile Include="DataStructures\Search\BinarySearch.cs" />
+    <Compile Include="DataStructures\Search\InterpolationSearch.cs" />
+    <Compile Include="DataStructures\Search\ISearch.cs" />
     <Compile Include="DataStructures\Search\LinearSearch.cs" />
+    <Compile Include="DataStructures\Sort\ISort.cs" />
     <Compile Include="DataStructures\Stack\ListStack.cs" />
     <Compile Include="DataStructures\Stack\VectorStack.cs" />
     <Compile Include="DataStructures\Stack\ArrayStack.cs" />
