+++ DSA/DSA.BLL/DataStructures/Search/ISearch.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DSA.BLL.DataStructures.Search
+{
+    public interface ISearch<T>
+    {
+        T Search(ICountable<T> array, T item);
+    }
+}
